Add possibility of invalid input to the loop
Assign the player choice functions to variables so that invalid input
can be detected inside the loop and the maximum number of rounds =
iterations will be increased.

This means:
The game should always have a total of 5 valid rounds. If the human
player enters an invalid value, this should NOT count as one of the 5
rounds. The player shall then be prompted again to input a value.