Add console log when user clicks "cancel"
Before this change, an error was thrown when clicking "cancel".

The console is supposed to print a message stating that the user
cancelled the game if "cancel" is clicked in the prompt window.

Another detail:
To prevent calling the function that prints empty scores at the end,
the humanChoice variable had to be declared outside of the "for loop"
scope (also, NOT a constant anymore).